# apps/backend/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"EventImageType": {
				"type": "string",
				"enum": [
					"flyer",
					"hero"
				]
			},
...
			"GetEventsPaginatedResponse": {
				"$ref": "#/components/schemas/PaginatedResponse__createdAt-Date--description-string-or-null--eventEndDate-Date--eventStartDate-Date--externalUrl-string--id-string--name-string--status-string--updatedAt-Date--venueAddress-string--venueName-string-or-null--lowestAvailableTicketPrice-number-or-null--lowestAvailableTicketCurrency-string-or-null--images_58__url-string--imageType-EventImageType_-Array__"
			},
...
			"SearchEventsResponse": {
				"items": {
					"properties": {
						"eventImages": {
							"items": {
								"properties": {
									"imageType": {
										"$ref": "#/components/schemas/EventImageType"
									},
									"url": {
										"type": "string"
									}
								},
								"required": [
									"imageType",
									"url"
								],
								"type": "object"
							},
							"type": "array"
						},
						"venueName": {
							"type": "string",
							"nullable": true
						},
						"venueAddress": {
							"type": "string"
						},
						"updatedAt": {
							"type": "string",
							"format": "date-time"
						},
						"status": {
							"type": "string"
						},
						"name": {
							"type": "string"
						},
						"id": {
							"type": "string"
						},
						"externalUrl": {
							"type": "string"
						},
						"eventStartDate": {
							"type": "string",
							"format": "date-time"
						},
						"eventEndDate": {
							"type": "string",
							"format": "date-time"
						},
						"description": {
							"type": "string",
							"nullable": true
						},
						"createdAt": {
							"type": "string",
							"format": "date-time"
						}
					},
					"required": [
						"eventImages",
						"venueName",
						"venueAddress",
						"updatedAt",
						"status",
						"name",
						"id",
						"externalUrl",
						"eventStartDate",
						"eventEndDate",
						"description",
						"createdAt"
					],
					"type": "object"
				},
				"type": "array"
			},
			"EventTicketCurrency": {
				"type": "string",
				"enum": [
					"USD",
					"UYU"
				]
			},
			"GetEventByIdResponse": {
				"properties": {
					"ticketWaves": {
						"items": {
							"properties": {
								"priceGroups": {
									"items": {
										"properties": {
											"availableTickets": {
												"anyOf": [
													{
														"type": "string"
													},
													{
														"type": "number",
														"format": "double"
													}
												]
											},
											"price": {
												"type": "string"
											}
										},
										"required": [
											"availableTickets",
											"price"
										],
										"type": "object"
									},
									"type": "array"
								},
								"faceValue": {
									"type": "string"
								},
								"currency": {
									"$ref": "#/components/schemas/EventTicketCurrency"
								},
								"name": {
									"type": "string"
								},
								"id": {
									"type": "string"
								},
								"description": {
									"type": "string",
									"nullable": true
								}
							},
							"required": [
								"priceGroups",
								"faceValue",
								"currency",
								"name",
								"id",
								"description"
							],
							"type": "object"
						},
						"type": "array"
					},
					"eventImages": {
						"items": {
							"properties": {
								"imageType": {
									"$ref": "#/components/schemas/EventImageType"
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"imageType",
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"venueName": {
						"type": "string",
						"nullable": true
					},
					"venueAddress": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"externalUrl": {
						"type": "string"
					},
					"eventStartDate": {
						"type": "string",
						"format": "date-time"
					},
					"eventEndDate": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"ticketWaves",
					"eventImages",
					"venueName",
					"venueAddress",
					"updatedAt",
					"status",
					"name",
					"id",
					"externalUrl",
					"eventStartDate",
					"eventEndDate",
					"description",
					"createdAt"
				],
				"type": "object"
			},
			"Record_string.any_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"CreateTicketListingResponse": {
				"properties": {
					"ticketWaveId": {
						"type": "string"
					},
					"soldAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"publisherUserId": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"id": {
						"type": "string"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"listingTickets": {
						"items": {
							"properties": {
								"ticketNumber": {
									"type": "number",
									"format": "double"
								},
								"price": {
									"type": "string"
								},
								"listingId": {
									"type": "string"
								},
								"documentUploadRequiredNotifiedAt": {
									"type": "string",
									"format": "date-time",
									"nullable": true
								},
								"documentUploadedAt": {
									"type": "string",
									"format": "date-time",
									"nullable": true
								},
								"documentType": {
									"type": "string",
									"nullable": true
								},
								"documentSizeBytes": {
									"type": "number",
									"format": "double",
									"nullable": true
								},
								"documentPath": {
									"type": "string",
									"nullable": true
								},
								"documentOriginalName": {
									"type": "string",
									"nullable": true
								},
								"cancelledAt": {
									"type": "string",
									"format": "date-time",
									"nullable": true
								},
								"soldAt": {
									"type": "string",
									"format": "date-time",
									"nullable": true
								},
								"updatedAt": {
									"type": "string",
									"format": "date-time"
								},
								"id": {
									"type": "string"
								},
								"deletedAt": {
									"type": "string",
									"format": "date-time",
									"nullable": true
								},
								"createdAt": {
									"type": "string",
									"format": "date-time"
								}
							},
							"required": [
								"ticketNumber",
								"price",
								"listingId",
								"documentUploadRequiredNotifiedAt",
								"documentUploadedAt",
								"documentType",
								"documentSizeBytes",
								"documentPath",
								"documentOriginalName",
								"cancelledAt",
								"soldAt",
								"updatedAt",
								"id",
								"deletedAt",
								"createdAt"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"ticketWaveId",
					"soldAt",
					"publisherUserId",
					"updatedAt",
					"id",
					"deletedAt",
					"createdAt",
					"listingTickets"
				],
				"type": "object"
			},
			"UnauthorizedError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"isOperational": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"message",
					"statusCode",
					"isOperational"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BadRequestError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"isOperational": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"message",
					"statusCode",
					"isOperational"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NotFoundError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"isOperational": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"message",
					"statusCode",
					"isOperational"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidationError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"isOperational": {
						"type": "boolean"
					},
					"metadata": {
						"$ref": "#/components/schemas/Record_string.any_"
					}
				},
				"required": [
					"name",
					"message",
					"statusCode",
					"isOperational"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTicketListingRouteBody": {
				"properties": {
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"ticketWaveId": {
						"type": "string"
					},
					"eventId": {
						"type": "string"
					}
				},
				"required": [
					"quantity",
					"price",
					"ticketWaveId",
					"eventId"
				],
				"type": "object"
			},
...
			"UploadDocumentResponse": {
				"properties": {
					"documentUrl": {
						"type": "string"
					},
					"document": {
						"properties": {
							"version": {
								"type": "number",
								"format": "double"
							},
							"verifiedBy": {
								"type": "string",
								"nullable": true
							},
							"verifiedAt": {
								"type": "string",
								"format": "date-time",
								"nullable": true
							},
							"uploadedAt": {
								"type": "string",
								"format": "date-time"
							},
							"ticketId": {
								"type": "string"
							},
							"storagePath": {
								"type": "string"
							},
							"sizeBytes": {
								"type": "number",
								"format": "double"
							},
							"originalName": {
								"type": "string"
							},
							"mimeType": {
								"type": "string"
							},
							"isPrimary": {
								"type": "boolean"
							},
							"fileName": {
								"type": "string"
							},
							"documentType": {
								"type": "string"
							},
							"updatedAt": {
								"type": "string",
								"format": "date-time"
							},
							"status": {
								"type": "string"
							},
							"id": {
								"type": "string"
							},
							"deletedAt": {
								"type": "string",
								"format": "date-time",
								"nullable": true
							},
							"createdAt": {
								"type": "string",
								"format": "date-time"
							}
						},
						"required": [
							"version",
							"verifiedBy",
							"verifiedAt",
							"uploadedAt",
							"ticketId",
							"storagePath",
							"sizeBytes",
							"originalName",
							"mimeType",
							"isPrimary",
							"fileName",
							"documentType",
							"updatedAt",
							"status",
							"id",
							"deletedAt",
							"createdAt"
						],
						"type": "object"
					}
				},
				"required": [
					"documentUrl"
				],
				"type": "object"
			},
...
			"Record_string.Record_string.number__": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"CreateOrderRouteBody": {
				"properties": {
					"ticketSelections": {
						"$ref": "#/components/schemas/Record_string.Record_string.number__"
					},
					"eventId": {
						"type": "string"
					}
				},
				"required": [
					"ticketSelections",
					"eventId"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@revendiste/backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/events": {
			"get": {
				"operationId": "GetAllPaginated",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetEventsPaginatedResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/events/search": {
			"get": {
				"operationId": "GetBySearch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SearchEventsResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/events/{eventId}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetEventByIdResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/ticket-listings": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateTicketListingResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request body, missing required fields, or invalid data types",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestError"
								}
							}
						}
					},
					"401": {
						"description": "Authentication required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"404": {
						"description": "Event not found or ticket wave not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed: Cannot create listing for finished event, ticket wave does not belong to event, price exceeds face value, or quantity must be greater than 0",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					}
				},
				"tags": [
					"Ticket Listings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTicketListingRouteBody"
							}
						}
					}
				}
			}
		},
		"/ticket-listings/my-listings": {
			"get": {
				"operationId": "GetMyListings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserListingsResponse"
								}
							}
						}
					},
					"401": {
						"description": "Authentication required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					}
				},
				"tags": [
					"Ticket Listings"
				],
				"security": [],
				"parameters": []
			}
		},
		"/ticket-listings/tickets/{ticketId}/document": {
			"post": {
				"operationId": "UploadDocument",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UploadDocumentResponse"
								}
							}
						}
					},
					"400": {
						"description": "No file uploaded or invalid file",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestError"
								}
							}
						}
					},
					"401": {
						"description": "Authentication required or not authorized to upload for this ticket",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"404": {
						"description": "Ticket not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed: ticket not sold, event has ended, invalid file type, or file too large",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					}
				},
				"description": "Upload a ticket document\n\nSellers can upload PDF/image documents for their sold tickets.\nDocuments must be uploaded before the event ends.",
				"tags": [
					"Ticket Listings"
				],
				"security": [],
				"parameters": [
					{
						"description": "- ID of the ticket",
						"in": "path",
						"name": "ticketId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary",
										"description": "- Document file (PDF, PNG, JPG, JPEG - max 10MB)"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateDocument",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UploadDocumentResponse"
								}
							}
						}
					},
					"400": {
						"description": "No file uploaded or invalid file",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestError"
								}
							}
						}
					},
					"401": {
						"description": "Authentication required or not authorized to upload for this ticket",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"404": {
						"description": "Ticket not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed: ticket not sold, event has ended, invalid file type, or file too large",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					}
				},
				"description": "Update/replace a ticket document\n\nUploads a new version of the document. Previous versions are kept for audit trail.\nCan only update before the event ends.",
				"tags": [
					"Ticket Listings"
				],
				"security": [],
				"parameters": [
					{
						"description": "- ID of the ticket",
						"in": "path",
						"name": "ticketId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary",
										"description": "- New document file (PDF, PNG, JPG, JPEG - max 10MB)"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/orders": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateOrderResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request body, missing required fields, or invalid data types",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestError"
								}
							}
						}
					},
					"401": {
						"description": "Authentication required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"404": {
						"description": "Event not found or ticket wave not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					},
					"422": {
						"description": "Validation failed: Not enough tickets available, event has ended, or invalid ticket selection",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationError"
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOrderRouteBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetMyOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserOrdersResponse"
								}
							}
						}
					},
					"401": {
						"description": "Authentication required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": []
			}
		},
		"/orders/{orderId}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetOrderByIdResponse"
								}
							}
						}
					},
					"401": {
						"description": "Authentication required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"404": {
						"description": "Order not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/orders/{orderId}/tickets": {
			"get": {
				"operationId": "GetOrderTickets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetOrderTicketsResponse"
								}
							}
						}
					},
					"401": {
						"description": "Authentication required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Not authorized to view tickets for this order",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"404": {
						"description": "Order not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundError"
								}
							}
						}
					}
				},
				"tags": [
					"Orders"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "orderId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}